Men's decathlon performances

A dataset containing the performances of 23 athletes in the men's decathlon at the 2016 Summer Olympics.

dec

Format

A data frame with 23 rows and 24 variables. The variables `100m`, LJ, SP, HJ, `400m`, `110mH`, DT, PV, JT, and `1500m` correspond to the performances of the athletes for the ten events comprising the decathlon. Those variables ending in _p (e.g., `100m_p`) represent the points athletes earn for their performances in each of the ten events. A full description of the 24 variables is below.

rank

rank of athlete

athlete

name of athlete

nationality

nationality of athlete

score_total

overall score

100m

100m result, in seconds

100m_p

100m points

LJ

long jump result, in meters

LJ_p

long jump points

SP

shot put result, in meters

SP_p

shot put points

HJ

high jump result, in meters

HJ_p

high jump points

400m

400m result, in seconds

400m_p

400m points

110mH

110m hurdles result, in seconds

110mH_p

110m hurdles points

DT

discus throw result, in meters

DT_p

discus throw points

PV

pole vault result, in meters

PV_p

pole vault points

JT

javelin throw result, in meters

JT_p

javelin throw points

1500m

1500m result, in minutes

1500m_p

1500m points

Source

https://en.wikipedia.org/wiki/Athletics_at_the_2016_Summer_Olympics_–_Men%27s_decathlon